+++ CHANGELOG.md
@@ -7,6 +7,11 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ## x.y.z
 
 ### Added
+- Add an HtmlNormalizer class
+  ([#513](https://github.com/jjriv/emogrifier/pull/513),
+  [#516](https://github.com/jjriv/emogrifier/pull/516))
+- Add a CssInliner class
+  ([#514](https://github.com/jjriv/emogrifier/pull/514))
 - Composer scripts for the various CI build steps
 - Validate the composer.json on Travis
   ([#476](https://github.com/jjriv/emogrifier/pull/476))